requirements.txt
pandas>=1.3.0
openpyxl>=3.0.7
openai>=1.0.0
python-dotenv>=0.19.0
